XRDDEV-562 Implement ownerChange management request client on Security Server
- Add new management request code ownerChange.
- Implement management request builder for ownerChange.
- Implement management request sender for ownerChange.
- Refactor ClientRegRequest - implement new GenericClientRequest class that has two child classes that share the same implementation.
