+++ src/common-ui/src/main/java/ee/ria/xroad/common/request/ManagementRequestBuilder.java
@@ -45,6 +45,7 @@ import static ee.ria.xroad.common.request.ManagementRequests.AUTH_CERT_DELETION;
 import static ee.ria.xroad.common.request.ManagementRequests.AUTH_CERT_REG;
 import static ee.ria.xroad.common.request.ManagementRequests.CLIENT_DELETION;
 import static ee.ria.xroad.common.request.ManagementRequests.CLIENT_REG;
+import static ee.ria.xroad.common.request.ManagementRequests.OWNER_CHANGE;
 
 @Slf4j
 final class ManagementRequestBuilder {
@@ -101,6 +102,16 @@ final class ManagementRequestBuilder {
         request.setClient(client);
 
         return buildMessage(element(CLIENT_DELETION, ClientRequestType.class, request));
+    }
+
+    SoapMessageImpl buildOwnerChangeRequest(SecurityServerId securityServer, ClientId client) throws Exception {
+        log.debug("buildOwnerChangeRequest(server: {}, client: {})", securityServer, client);
+
+        ClientRequestType request = FACTORY.createClientRequestType();
+        request.setServer(securityServer);
+        request.setClient(client);
+
+        return buildMessage(element(OWNER_CHANGE, ClientRequestType.class, request));
     }
 
     // -- Private helper methods ----------------------------------------------
